# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /action/{action_id}/subscription/elements:
    get:
      summary: "get the subscription elements for an action"
      tags:
        - "action"
        - "subscription"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "the subscription elements"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  type: object
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
        "400":
          description: "if the action doesn't have a subscription"
          schema:
            $ref: "#/definitions/ErrorResult"
...
  /dataset/{dataset_id}/subscription:
    post:
      summary: "create a subscription to a dataset"
      parameters:
        - name: "dataset_id"
          type: "string"
          in: "path"
          required: true
        - name: "subscription"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Subscription"
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResult"
...
  /subscription:
    get:
      summary: "find subscriptions"
      tags:
        - "subscription"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "the subscriptions"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: "#/definitions/Subscription"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
  /subscription/{subscription_id}:
    parameters:
      - name: subscription_id
        in: path
        type: string
        required: true
    get:
      summary: "get a subscription"
      tags:
        - "subscription"
      responses:
        "200":
          description: "the subscription"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Subscription"
    put:
      summary: "update a subscription"
      tags:
        - "subscription"
      parameters:
        - name: subscription
          in: body
          required: true
          schema:
            $ref: "#/definitions/Subscription"
      responses:
        "200":
          description: "the subscription"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Subscription"
        "400":
          description: "subscription is in an invalid state to update"
          schema:
            $ref: "#/definitions/ErrorResult"
    patch:
      summary: "update a subscription using a JSON Patch"
      tags:
        - "subscription"
      parameters:
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: "the subscription"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Subscription"
        "400":
          description: "subscription is in an invalid state to update"
          schema:
            $ref: "#/definitions/ErrorResult"
    delete:
      summary: "delete a subscription"
      tags:
        - "subscription"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/OKResult"
  /subscription/{subscription_id}/element_stats:
    get:
      summary: "get the element statistics for a subscription"
      tags:
        - "subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
      responses:
        "200":
          description: "the element statistics"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  type: object
  /subscription/{subscription_id}/elements:
    get:
      summary: "get the elements for a subscription"
      tags:
        - "subscription"
      parameters:
        - name: subscription_id
          in: path
          type: string
          required: true
        - name: state
          in: query
          type: string
        - name: processed_after_s3_path
          in: query
          type: string
        - name: limit
          in: query
          type: integer
        - name: offset
          in: query
          type: integer
      responses:
        "200":
          description: "the subscription elements"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  type: object
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
  /trigger:
    post:
      summary: "create a new trigger"
      tags:
        - "trigger"
      parameters:
        - name: trigger
          in: body
          required: true
          schema:
            $ref: "#/definitions/Trigger"
      responses:
        "200":
          description: "the trigger"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Trigger"
    get:
      summary: "find triggers"
      tags:
        - "trigger"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "the triggers"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: "#/definitions/Trigger"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
  /trigger/{trigger_id}:
    parameters:
      - name: trigger_id
        in: path
        type: string
        required: true
    get:
      summary: "get a trigger"
      tags:
        - "trigger"
      responses:
        "200":
          description: "the trigger"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Trigger"
    put:
      summary: "update a trigger"
      tags:
        - "trigger"
      parameters:
        - name: trigger
          in: body
          required: true
          schema:
            $ref: "#/definitions/Trigger"
      responses:
        "200":
          description: "the trigger"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Trigger"
    patch:
      summary: "patch a trigger"
      tags:
        - "trigger"
      parameters:
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: "the trigger"
          schema:
            type: object
            properties:
              results:
                $ref: "#/definitions/Trigger"
    delete:
      summary: "delete a trigger"
      tags:
        - "trigger"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/OKResult"
  /trigger_type:
    get:
      summary: "get trigger types"
      tags:
        - "trigger"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
      responses:
        "200":
          description: "the trigger types"
          schema:
            type: object
            properties:
              results:
                type: array
                items:
                  $ref: "#/definitions/TriggerType"
              limit:
                type: "integer"
              offset:
                type: "integer"
              total:
                type: "integer"
...
parameters:
  limitParam:
    name: "limit"
    type: "integer"
    in: "query"
    description: |
      Specify the maximum number of records that should be returned by this request.
  offsetParam:
    name: "offset"
    type: "integer"
    in: "query"
    description: |
      Specify the starting index for the next chunk of results to be returned by this request.
  filtersParam:
    name: filters
    in: query
    type: string
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of Filter objects (as defined in
      the definitions section).
...
  patchParam:
    name: "patch"
    description: "a JSON Patch object"
    in: "body"
    required: true
    schema:
      $ref: '#/definitions/JSONPatch'
#
# Definitions are auto-generated by dart.tool.export_swagger_definitions
# Do not manually edit!
#
definitions:
...
  ErrorResult:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
...
  JSONPatch:
    description: JSON Patch object as defined by http://json.schemastore.org/json-patch
    type: object
...
  OKResult:
    properties:
      results:
        enum:
        - OK
        type: string
    type: object
...
  Subscription:
    additionalProperties: false
    properties:
      created:
        type: string
      data:
        $ref: '#/definitions/SubscriptionData'
      id:
        type: string
      updated:
        type: string
      version_id:
        type: integer
    required:
    - data
    type: object
  SubscriptionData:
    additionalProperties: false
    properties:
      dataset_id:
        type: string
      failed_time:
        type: string
      generating_time:
        type: string
      initial_active_time:
        type: string
      message_id:
        type: string
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      on_failure_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      on_success_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      queued_time:
        type: string
      s3_path_end_prefix_exclusive:
        description: The exclusive s3 path end prefix
        pattern: ^s3://.+$
        type: string
      s3_path_regex_filter:
        description: A regex pattern the s3 path must match
        type: string
      s3_path_start_prefix_inclusive:
        description: The inclusive s3 path start prefix
        pattern: ^s3://.+$
        type: string
      state:
        default: INACTIVE
        enum:
        - INACTIVE
        - QUEUED
        - GENERATING
        - FAILED
        - ACTIVE
        type: string
      tags:
        default: []
        items:
          maxLength: 30
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
    required:
    - name
    - dataset_id
    type: object
  Trigger:
    additionalProperties: false
    properties:
      created:
        type: string
      data:
        $ref: '#/definitions/TriggerData'
      id:
        type: string
      updated:
        type: string
      version_id:
        type: integer
    required:
    - data
    type: object
  TriggerData:
    additionalProperties: false
    properties:
      args:
        type: object
      extra_data:
        type: object
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      state:
        default: INACTIVE
        enum:
        - ACTIVE
        - INACTIVE
        type: string
      tags:
        default: []
        items:
          maxLength: 30
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      trigger_type_name:
        type: string
      workflow_ids:
        default: []
        items:
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''workflow_id '' + $index }}'
          type: tabarray
    required:
    - name
    - trigger_type_name
    - args
    type: object
  TriggerType:
    properties:
      description:
        type: string
      name:
        type: string
      params_json_schema:
        type: object
    type: object
...